Revert "Optimize systemd dependency for containers"
It was generally a mistake to use `%systemd_ordering` in those
circumstances, as the tmpfiles stuff doesn't work and populate the
`/tmp` and `/run` directories. The reason why the packages install is
that all the commands in the macros have `|| :` at the end, so their
error state is ignored.

Something to re-consider for future attempts to fix/improve that:
If the tmpfiles and such were split out into a subpackage that'd be
required if and only if the kernel was installed or being installed,
that might work.

Related: rhbz#1654672
Related-Bug: #1804822
This reverts commit e07ed63d656e9d3f6688eea695b1ef8e08028993.